feat(ReservationApiControllerTest) : port 생성
